+++ merger.js
@@ -1,6 +1,5 @@
 #!/usr/bin/env node
 
-
 /*
  * Copyright (c) 2018-2019 João Pedro Martins Neves - All Rights Reserved.
  *
@@ -13,61 +12,55 @@
 const path = require('path');
 const chokidar = require('chokidar');
 const mergerCLI = require('./modules/mergerCLI');
-const config = require('./modules/config');
+const Config = require('./modules/config');
 const async = require( './node_modules/neo-async' );
 const selectSourceFile = require('./modules/CLIModules/selectSourceFilePrompt');
 const parseImports = require( './modules/buildModules/parseImports' );
 const build = require( './modules/buildModules/build' );
-const sourceFileModel = require( './models/sourceFileModel' );
 
 // #region PROGRAM
 
-mergerCLI( ( newConfig ) => {
-  config( newConfig, () => {
+mergerCLI( async ( newConfig ) => {
+  await Config.init( newConfig );
+  let sourceFiles = await selectSourceFile();
 
-    selectSourceFile( ( sourceFile ) => {
-      /** @type { sourceFileModel[] } */
-      let files = [sourceFile];
+  // If "sourceFiles" is Array it means that the user chose the "All" (sourceFiles) option in selectSourceFile().
+  if ( Array.isArray( sourceFiles ) ) {
+    // Execute a one time build only.
+    global.config.autoBuild = false;
 
-      // If "sourceFile" is Array it means that the user chose the "All" (files) option in selectSourceFile().
-      if (Array.isArray(sourceFile)) {
-        files = sourceFile;
-        // Execute a one time build only.
-        global.config.autoBuild = false;
-      }
+  } else {
+    sourceFiles = [sourceFiles];
+  }
 
-      async.eachSeries( files, ( file, Callback ) => {
+  async.eachSeries( sourceFiles, async ( sourceFile, Callback ) => {
+    await Config.setCustomConfig( sourceFile.source );
+    const buildOrder = await parseImports( sourceFile.source );
 
-        parseImports( file.source, async ( buildOrder ) => {
-          // Execute one time builds:
-          if ( !global.config.autoBuild ) {
-            await build( file, buildOrder );
-            return Callback();
-          }
+    // Execute one time builds:
+    if ( !global.config.autoBuild ) {
+      await build( sourceFile, buildOrder );
+      return Callback();
 
-          // Execute an auto builds (with file watcher):
-          else {
-            const whatcher = chokidar.watch( buildOrder, { persistent: true, cwd: path.dirname( file.source ) } );
+      // Execute an auto build session (with file watcher):
+    } else {
+      const whatcher = chokidar.watch( buildOrder, { persistent: true, cwd: path.dirname( sourceFile.source ) } );
+      whatcher
+        .on( 'ready', async () => {
+          console.info( ' Inicial scan complete. Ready to build on changes...' );
+          await build( sourceFile, buildOrder );
+          return Callback();
+        } )
+        .on( 'error', err => console.error( 'Auto build error: ', err ) )
+        .on( 'change', async ( path, stats ) => {
+          await build( sourceFile, null );
+        } );
+    }
 
-            whatcher
-              .on( 'ready', async () => {
-                console.info( ' Inicial scan complete. Ready to build on changes...' );
-                await build( file, buildOrder );
-                return Callback();
-              })
-              .on('error', err => console.error('Auto build error: ', err))
-              .on( 'change', async ( path, stats ) => {
-                await build(file, null);
-            });
-          }
-        });
+  }, ( err ) => {
+    if ( err ) throw err;
+  } );
 
-      }, (err) => {
-          if (err) throw err;
-      });
-
-    });
-  });
 } );
 
 // #endregion