+++ modules/CLIModules/addFilesPrompt.js
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018-2019 João Pedro Martins Neves - All Rights Reserved.
+ * Copyright (c) 2018-2019 JoÃ£o Pedro Martins Neves - All Rights Reserved.
  *
  * MergerJS (merger-js) is licensed under the MIT license, located in
  * the root of this project, under the name "LICENSE.md".
@@ -9,8 +9,9 @@
 'use strict'
 const path = require('path');
 const prompt = require('../../node_modules/inquirer').createPromptModule();
-const SourceFile = require( '../../models/sourceFileModel' );
 const { readDir } = require( '../utils' );
+const SourceFile = require( '../../models/sourceFileModel' );
+const DEFAULT_BUILD_FILE_NAME = 'build.js';
 
 let questions = [
   {
@@ -36,7 +37,7 @@ let questions = [
     type: 'input',
     name: 'outputName',
     message: '\n Output file name.\n Default: ',
-    default: 'build.js'
+    default: DEFAULT_BUILD_FILE_NAME
   }
 ];
 
@@ -44,9 +45,11 @@ module.exports = async ( Callback ) => {
   const currentDir = process.cwd();
   /** @type { string[] } */
   const allFilesFromCurrentDir = await readDir( currentDir );
+
   for ( let i = 0; i < allFilesFromCurrentDir.length; ++i ) {
-    if ( path.extname( allFilesFromCurrentDir[i] ) !== '' )
+    if ( path.extname( allFilesFromCurrentDir[i] ) !== '' ) {
       questions[0].choices.push( allFilesFromCurrentDir[i] );
+    }
   }
 
   const sourceFile = new SourceFile();
@@ -61,8 +64,9 @@ module.exports = async ( Callback ) => {
     answers = await prompt( [questions[1]] );
 
     sourceFile.source = answers.customSourceFile;
-    if ( path.extname( sourceFile.source ) === '' )
+    if ( path.extname( sourceFile.source ) === '' ) {
       sourceFile.source += '.js';
+    }
 
     sourceFile.source = path.join( currentDir, sourceFile.source );
   }
@@ -71,8 +75,13 @@ module.exports = async ( Callback ) => {
 
   sourceFile.output.path = path.join( currentDir, answers.outputPath );
   sourceFile.output.name = answers.outputName;
-  if ( path.extname( sourceFile.output.name ) === '' )
+
+  if ( sourceFile.output.name === '' ) {
+    sourceFile.output.name = DEFAULT_BUILD_FILE_NAME;
+
+  } else if ( path.extname( sourceFile.output.name ) === '' ) {
     sourceFile.output.name += '.js';
+  }
 
   return Callback( sourceFile );
 };