+++ tests/spec/mergerCLI_spec.js
@@ -1,0 +1,14 @@
+﻿/*
+ * Copyright (c) 2018-2019 João Pedro Martins Neves - All Rights Reserved.
+ *
+ * MergerJS (merger-js) is licensed under the MIT license, located in
+ * the root of this project, under the name "LICENSE.md".
+ *
+ */
+
+// Test merger log
+// Test merger set
+// Test merger add
+// Test merger rm
+// test merger build
+// Test the name of the build file